Simplify hockey example agent
Don't provide full package, just a single all-in-one script and a
requirements.txt.
