--- comprl-hockey-agent/pyproject.toml
@@ -1,28 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-
-[project]
-name = "comprl-hockey-agent"
-version = "1.0.0"
-description = "Example agent for the comprl hockey game."
-readme = "README.md"
-requires-python = ">=3.10"
-dependencies = [
-    "comprl",
-    "numpy",
-    "hockey @ git+https://git@github.com/martius-lab/laser-hockey-env.git@cc48daa135e2cca3e620b90431e57870f61d043c",
-]
-
-[project.scripts]
-comprl-hockey-agent = "comprl_hockey_agent:main"
-
-
-[tool.hatch.metadata]
-allow-direct-references = true
-
-
-[tool.mypy]
-exclude = ["build"]
-ignore_missing_imports = true
