--- comprl-hockey-agent/src/comprl_hockey_agent/__init__.py
@@ -1,44 +1,0 @@
-from __future__ import annotations
-
-import argparse
-import typing
-
-from comprl.client import Agent, launch_client
-
-from .agents import HockeyAgent, RandomAgent
-
-
-# Function to initialize the agent.  This function is used with `launch_client` below,
-# to lauch the client and connect to the server.
-def initialize_agent(agent_args: list[str]) -> Agent:
-    # Use argparse to parse the arguments given in `agent_args`.
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--agent",
-        type=str,
-        choices=["weak", "strong", "random"],
-        default="weak",
-        help="Which agent to use.",
-    )
-    args = parser.parse_args(agent_args)
-
-    # Initialize the agent based on the arguments.
-    agent: Agent
-    if args.agent == "weak":
-        agent = HockeyAgent(weak=True)
-    elif args.agent == "strong":
-        agent = HockeyAgent(weak=False)
-    elif args.agent == "random":
-        agent = RandomAgent()
-    else:
-        raise ValueError(f"Unknown agent: {args.agent}")
-
-    # And finally return the agent.
-    return agent
-
-
-def main() -> None:
-    launch_client(initialize_agent)
-
-
-__all__: list[str] = []
